fix: operation cancel logging uncaught exception warning
The asyncio uncaught exception handler can be triggered when operations
are cancelled, causing an error to be printed to stderr.

AsyncioContextBase._on_done() schedules a cross-thread call to
Future.set_result, but doesn't check that the future is still not done
before the call later runs on the asyncio thread. This allows for an
operation to be cancelled in between the initial done() check and the
set_result() call, which then fails, triggering the uncaught exception
handler.

Now the callback re-checks if the Future is done before calling
set_result().